+++ robot-core/src/main/java/org/obolibrary/robot/Template.java
@@ -854,7 +854,6 @@ public class Template {
     // Instead of adding them in as we iterate through
     Map<Integer, Set<OWLClassExpression>> subclassExpressionColumns = new HashMap<>();
     Map<Integer, Set<OWLClassExpression>> equivalentExpressionColumns = new HashMap<>();
-    Map<Integer, Set<OWLClassExpression>> intersectionEquivalentExpressionColumns = new HashMap<>();
     Map<Integer, Set<OWLClassExpression>> disjointExpressionColumns = new HashMap<>();
     for (int column = 0; column < templates.size(); column++) {
       String template = templates.get(column);
@@ -903,7 +902,7 @@ public class Template {
                     name, checker, parser, template, value, rowNum, column));
             break;
           case "equivalent":
-            intersectionEquivalentExpressionColumns.put(
+            equivalentExpressionColumns.put(
                 column,
                 TemplateHelper.getClassExpressions(
                     name, checker, parser, template, value, rowNum, column));
@@ -926,15 +925,6 @@ public class Template {
     }
     if (!equivalentExpressionColumns.isEmpty()) {
       addEquivalentClassesAxioms(cls, equivalentExpressionColumns, row);
-    }
-    if (!intersectionEquivalentExpressionColumns.isEmpty()) {
-      // Special case to support legacy "C"/"equivalent" class type
-      // Which is the intersection of all C columns
-      if (intersectionEquivalentExpressionColumns.size() == 1) {
-        addEquivalentClassesAxioms(cls, intersectionEquivalentExpressionColumns, row);
-      } else {
-        addIntersectionEquivalentClassesAxioms(cls, intersectionEquivalentExpressionColumns, row);
-      }
     }
     if (!disjointExpressionColumns.isEmpty()) {
       addDisjointClassAxioms(cls, disjointExpressionColumns, row);
@@ -967,9 +957,9 @@ public class Template {
   }
 
   /**
-   * Given an OWLClass, a map of column number to class expressions, and the row containing the
-   * class details, generate equivalent class axioms for the class where the equivalents are the
-   * class expressions. Maybe annotate the axioms.
+   * Given an OWLClass, a map of column number to class expressions, and the row for this class,
+   * generate an equivalentClasses axiom for the class where the equivalent is the intersection of
+   * the provided class expressions. Maybe annotate the axioms.
    *
    * @param cls OWLClass to create equivalentClasses axiom for
    * @param expressionColumns map of column number to equivalent class expression
@@ -981,41 +971,23 @@ public class Template {
       throws Exception {
     Set<OWLAnnotation> axiomAnnotations = new HashSet<>();
     Set<OWLClassExpression> expressions = new HashSet<>();
-    expressions.add(cls);
     for (int column : expressionColumns.keySet()) {
       // Maybe get an annotation on the expression (all annotations will be on the one intersection)
       axiomAnnotations.addAll(maybeGetAxiomAnnotations(row, column));
       // Add all expressions to the set of expressions
       expressions.addAll(expressionColumns.get(column));
     }
-    // Create the axiom as an intersection of the provided expressions
-    axioms.add(dataFactory.getOWLEquivalentClassesAxiom(expressions, axiomAnnotations));
-  }
 
-  /**
-   * Given an OWLClass, a map of column number to class expressions, and the row for this class,
-   * generate an equivalentClasses axiom for the class where the equivalent is the intersection of
-   * the provided class expressions. Maybe annotate the axioms.
-   *
-   * @param cls OWLClass to create equivalentClasses axiom for
-   * @param expressionColumns map of column number to equivalent class expression
-   * @param row list of template values for given class
-   * @throws Exception on issue getting axiom annotations
-   */
-  private void addIntersectionEquivalentClassesAxioms(
-      OWLClass cls, Map<Integer, Set<OWLClassExpression>> expressionColumns, List<String> row)
-      throws Exception {
-    Set<OWLAnnotation> axiomAnnotations = new HashSet<>();
-    Set<OWLClassExpression> expressions = new HashSet<>();
-    for (int column : expressionColumns.keySet()) {
-      // Maybe get an annotation on the expression (all annotations will be on the one intersection)
-      axiomAnnotations.addAll(maybeGetAxiomAnnotations(row, column));
-      // Add all expressions to the set of expressions
-      expressions.addAll(expressionColumns.get(column));
+    if (expressions.size() == 1) {
+      // Create a single equivalent class axiom (no intersection)
+      axioms.add(
+          dataFactory.getOWLEquivalentClassesAxiom(
+              cls, expressions.iterator().next(), axiomAnnotations));
+    } else if (expressions.size() > 1) {
+      // Create the axiom as an intersection of the provided expressions
+      OWLObjectIntersectionOf intersection = dataFactory.getOWLObjectIntersectionOf(expressions);
+      axioms.add(dataFactory.getOWLEquivalentClassesAxiom(cls, intersection, axiomAnnotations));
     }
-    // Create the axiom as an intersection of the provided expressions
-    OWLObjectIntersectionOf intersection = dataFactory.getOWLObjectIntersectionOf(expressions);
-    axioms.add(dataFactory.getOWLEquivalentClassesAxiom(cls, intersection, axiomAnnotations));
   }
 
   /**